pyproject.toml
[tool.poetry]
name = "guru_test"
version = "0.1.0"
description = ""
authors = ["Fennr <fenrir1121@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
DateTime = "^4.7"
django-filter = "^22.1"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
